+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/pages/employee/Login/index.js
@@ -4,7 +4,7 @@ import { Route, Switch, useRouteMatch } from "react-router-dom";
 import { loginConfig as defaultLoginConfig } from "./config";
 import { LoginOtpConfig as defaultLoginOtpConfig } from "./ConfigOtp";
 import LoginComponent from "./login";
-import { useHistory, useLocation } from "react-router-dom";
+import { useLoginConfig } from "../../../hooks/useLoginConfig";
 
 const EmployeeLogin = ({ stateCode }) => {
   const { t } = useTranslation();
@@ -38,15 +38,8 @@ const EmployeeLogin = ({ stateCode }) => {
     modulePrefix,
   });
 
-  const { data: mdmsData, isLoading } = Digit.Hooks.useCommonMDMS(stateCode, "commonUiConfig", ["LoginConfig"], {
-    select: (data) => {
-      return {
-        config: data?.commonUiConfig?.LoginConfig,
-      };
-    },
-    retry: false,
-  });
-
+  const { data : mdmsData, isLoading } = useLoginConfig(stateCode)
+  
   //let loginConfig = mdmsData?.config ? mdmsData?.config : defaultLoginConfig;
   useEffect(() => {
     if (isLoading == false && mdmsData?.config) {