+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/pages/employee/index.js
@@ -93,7 +93,7 @@ const EmployeeApp = ({
                 <Otp isLogin={true} />
               </Route>
               <Route path={`${path}/user/forgot-password`}>
-                <ForgotPassword />
+                <ForgotPassword stateCode={stateCode}/>
               </Route>
               <Route path={`${path}/user/change-password`}>
                 <ChangePassword />